requirements.txt
# 핵심 의존성
python-dotenv>=1.0.0
click>=8.1.6
rich>=13.4.2
jinja2>=3.1.2

# 웹 요청 및 파싱
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3

# 설정 및 데이터 처리
PyYAML>=6.0.1

# 개발 및 테스트용
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0

# 워드프레스 연동용
python-wordpress-xmlrpc>=2.3

# 로깅 및 모니터링
colorlog>=6.7.0
psutil>=5.9.5

# 보안
cryptography>=41.0.3

# 데이터 처리 (Python 3.13 호환 버전)
pandas>=2.1.0
openpyxl>=3.1.2

# 날짜/시간 처리
python-dateutil>=2.8.2
